Move namespace to org.janusgraph
Solves #10
