+++ conf/janusgraph-foundationdb-es-server.properties
@@ -15,7 +15,7 @@ gremlin.graph=org.janusgraph.core.JanusGraphFactory
 # Default:    (no default value)
 # Data Type:  String
 # Mutability: LOCAL
-storage.backend=com.experoinc.janusgraph.diskstorage.foundationdb.FoundationDBStoreManager
+storage.backend=org.janusgraph.diskstorage.foundationdb.FoundationDBStoreManager
 
 # Whether to enable JanusGraph's database-level cache, which is shared
 # across all transactions. Enabling this option speeds up traversals by
