--- src/test/java/com/experoinc/janusgraph/graphdb/foundationdb/FoundationDBGraphPerformanceMemoryTest.java
@@ -1,36 +1,0 @@
-// Copyright 2018 Expero Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.experoinc.janusgraph.graphdb.foundationdb;
-
-import org.janusgraph.diskstorage.configuration.WriteConfiguration;
-import org.janusgraph.graphdb.JanusGraphPerformanceMemoryTest;
-import com.experoinc.janusgraph.FoundationDBContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-/**
- * @author Ted Wilmes (twilmes@gmail.com)
- */
-@Testcontainers
-public class FoundationDBGraphPerformanceMemoryTest extends JanusGraphPerformanceMemoryTest {
-
-    @Container
-    public static FoundationDBContainer container = new FoundationDBContainer();
-
-    @Override
-    public WriteConfiguration getConfiguration() {
-        return container.getFoundationDBGraphConfiguration();
-    }
-}
