+++ src/test/java/org/janusgraph/blueprints/structure/FoundationDBJanusGraphStructureTest.java
@@ -1,0 +1,30 @@
+// Copyright 2018 Expero Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.janusgraph.blueprints.structure;
+
+import org.apache.tinkerpop.gremlin.GraphProviderClass;
+import org.apache.tinkerpop.gremlin.structure.StructureStandardSuite;
+import org.janusgraph.core.JanusGraph;
+import org.junit.runner.RunWith;
+
+import org.janusgraph.blueprints.FoundationDBGraphProvider;
+
+/**
+ * @author Ted Wilmes (twilmes@gmail.com)
+ */
+@RunWith(StructureStandardSuite.class)
+@GraphProviderClass(provider = FoundationDBGraphProvider.class, graph = JanusGraph.class)
+public class FoundationDBJanusGraphStructureTest {
+}
